requirements.txt
pylint
flake8
click
black
polars-lts-cpu
numpy
matplotlib
pandas
pytest
